test(renderer): give the self-checks one fake window instead of three
